# hld/api/openapi.yaml
openapi: 3.1.0
info:
  title: HumanLayer Daemon REST API
  version: 1.0.0
  description: |
    REST API for HumanLayer daemon operations, providing session management,
    approval workflows, and real-time event streaming capabilities.
  contact:
    name: HumanLayer Support
    url: https://humanlayer.dev
servers:
  - url: http://localhost:7777/api/v1
    description: Local daemon server
paths:
...
  /stream/events:
    get:
      operationId: streamEvents
      summary: Server-Sent Events stream
      description: |
        Subscribe to real-time events using Server-Sent Events (SSE).
        This endpoint streams events as they occur in the system.

        **Note**: This endpoint uses Server-Sent Events which is not natively
        supported by OpenAPI 3.1. Client code generation will not work for
        this endpoint. Manual SSE client implementation is required using:
        - JavaScript/TypeScript: Native EventSource API
        - Go: r3labs/sse or similar SSE client library
        - Other languages: Language-specific SSE client libraries
      tags:
        - Events
      x-sse: true
      parameters:
        - name: eventTypes
          in: query
          description: Filter by event types
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/EventType'
        - name: sessionId
          in: query
          description: Filter events by session ID
          schema:
            type: string
        - name: runId
          in: query
          description: Filter events by run ID
          schema:
            type: string
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Server-Sent Events stream. Each event follows the format:
                  data: {"type": "event_type", "timestamp": "ISO8601", "data": {...}}

                  Keepalive messages are sent every 30 seconds:
                  : keepalive
              example: |
                data: {"type":"new_approval","timestamp":"2024-01-01T12:00:00Z","data":{"approval_id":"appr_123","session_id":"sess_456","tool_name":"execute_command"}}

                : keepalive

                data: {"type":"session_status_changed","timestamp":"2024-01-01T12:00:01Z","data":{"session_id":"sess_456","old_status":"running","new_status":"completed"}}
components:
...
  schemas:
...
    EventType:
      type: string
      enum:
        - new_approval
        - approval_resolved
        - session_status_changed
        - conversation_updated
      description: Type of system event
...